requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
pytest-watch==4.2.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.1
tzdata==2024.1
watchdog==4.0.1
